Move backup and restore logics to backup-logic module
