+++ shared/cryptography/src/main/java/ir/mehdiyari/krypt/cryptography/utils/Base64Wrapper.kt
@@ -1,9 +1,5 @@
-package ir.mehdiyari.krypt.data.repositories
+package ir.mehdiyari.krypt.cryptography.utils
 
-import ir.mehdiyari.krypt.cryptography.utils.Base64
-import ir.mehdiyari.krypt.cryptography.utils.HashingUtils
-import ir.mehdiyari.krypt.cryptography.utils.SymmetricHelper
-import ir.mehdiyari.krypt.cryptography.utils.getBytesBetweenIndexes
 import javax.inject.Inject
 
 class Base64Wrapper @Inject constructor() {
