+++ shared/files/backup-logic/src/main/java/ir/mehdiyari/krypt/backup/logic/backup/DBBackupModel.kt
@@ -1,9 +1,9 @@
-package ir.mehdiyari.krypt.data.repositories.backup
+package ir.mehdiyari.krypt.backup.logic.backup
 
 import ir.mehdiyari.krypt.accounts.data.entity.AccountEntity
 import ir.mehdiyari.krypt.file.data.entity.FileEntity
 
-data class DBBackupModel(
+internal data class DBBackupModel(
     val account: AccountEntity,
     val files: List<FileEntity>
 )
