+++ shared/files/backup-logic/src/main/java/ir/mehdiyari/krypt/backup/logic/di/BackupModule.kt
@@ -1,0 +1,17 @@
+package ir.mehdiyari.krypt.backup.logic.di
+
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import ir.mehdiyari.krypt.backup.logic.backup.BackupRepository
+import ir.mehdiyari.krypt.backup.logic.backup.BackupRepositoryImpl
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal abstract class BackupModule {
+
+    @Binds
+    abstract fun bindBackupRepository(impl: BackupRepositoryImpl): BackupRepository
+
+}
