+++ shared/files/backup-logic/src/main/java/ir/mehdiyari/krypt/backup/logic/restore/RestoreRepositoryImpl.kt
@@ -1,6 +1,8 @@
-package ir.mehdiyari.krypt.data.repositories.restore
+package ir.mehdiyari.krypt.backup.logic.restore
 
 import ir.mehdiyari.krypt.accounts.data.dao.AccountsDao
+import ir.mehdiyari.krypt.backup.logic.backup.DBBackupModel
+import ir.mehdiyari.krypt.backup.logic.backup.DBBackupModelJsonAdapter
 import ir.mehdiyari.krypt.cryptography.api.ByteCryptography
 import ir.mehdiyari.krypt.cryptography.api.FileCryptography
 import ir.mehdiyari.krypt.cryptography.exceptions.DecryptException
@@ -10,8 +12,6 @@ import ir.mehdiyari.krypt.cryptography.utils.getAfterIndex
 import ir.mehdiyari.krypt.cryptography.utils.getBeforeIndex
 import ir.mehdiyari.krypt.cryptography.utils.getBestBufferSizeForFile
 import ir.mehdiyari.krypt.cryptography.utils.toLong
-import ir.mehdiyari.krypt.data.repositories.backup.DBBackupModel
-import ir.mehdiyari.krypt.data.repositories.backup.DBBackupModelJsonAdapter
 import ir.mehdiyari.krypt.file.data.dao.FilesDao
 import ir.mehdiyari.krypt.file.data.entity.FileTypeEnum
 import ir.mehdiyari.krypt.files.logic.repositories.utils.FilesUtilities
@@ -21,7 +21,7 @@ import java.io.FileNotFoundException
 import java.io.FileOutputStream
 import javax.crypto.SecretKey
 
-class RestoreRepositoryImpl(
+internal class RestoreRepositoryImpl(
     private val fileUtils: FilesUtilities,
     private val fileCryptography: FileCryptography,
     private val bytesCryptography: ByteCryptography,
